+++ package.json
@@ -17,8 +17,8 @@
     "test": "npm run test:unit && npm run test:integration",
     "prebuild": "npm run clean",
     "build": "tsc --project tsconfig.build.json  && npm run assets:copy",
-    "start": "npm run build && cd dist && node --require ./common/tracing.js ./index.js",
-    "start:dev": "npm run build && cd dist && node --enable-source-maps --require ./common/tracing.js ./index.js",
+    "start": "npm run build && cd dist && node --import ./instrumentation.mjs ./index.js",
+    "start:dev": "npm run build && cd dist && node --enable-source-maps --import ./instrumentation.mjs ./index.js",
     "assets:copy": "copyfiles -f ./config/* ./dist/config && copyfiles -f ./openapi3.yaml ./dist/ && copyfiles ./package.json dist",
     "clean": "rimraf dist",
     "install": "npx husky install"
@@ -48,7 +48,7 @@
     "@map-colonies/read-pkg": "0.0.1",
     "@map-colonies/telemetry": "^6.2.1",
     "@map-colonies/config": "https://ghatmpstorage.blob.core.windows.net/npm-packages/config-049b2c49e9573f5e1610fd9c11709b3e7498b90b.tgz",
-    "@map-colonies/schemas": "https://ghatmpstorage.blob.core.windows.net/npm-packages/schemas-44d2522224e8c1237756aca20cae0fdf9a3d8211.tgz",
+    "@map-colonies/schemas": "https://ghatmpstorage.blob.core.windows.net/npm-packages/schemas-b287ae5ac346cf62dc444419e70205fa7ed5acca.tgz",
     "@opentelemetry/api": "^1.7.0",
     "@opentelemetry/api-metrics": "0.23.0",
     "compression": "^1.7.4",
